+++ src/resources.c
@@ -12,6 +12,7 @@
 #include "resources/ota.h"
 #include "resources/actuators.h"
 #include "resources/sensors.h"
+#include "resources/policy.h"
 
 char resource_json_str[RESOURCE_JSON_STR_SIZE];
 char resource_e_json_str[RESOURCE_E_JSON_STR_SIZE];
@@ -32,6 +33,8 @@ const resource_t resources[] = {
     {"/inx/sensors/WallSwitch", NULL, &sensors_wallswitch_coap_handler},
     {"/inx/sensors/sensor1/context", NULL, &sensors_sensor_context_coap_handler},
     {"/inx/sensors/WallSwitch/context", NULL, &sensors_wallswitch_context_coap_handler},
+    {"/inx/policy", &policy_init, &policy_coap_handler},
+    {"/inx/policy/policy1", NULL, &policy_coap_handler},
     {NULL, NULL, NULL}
 };